Update base_connection.py function enable() to handle no password
Change line 1044 to 1047 to not send the enable secret password when it doesn't have an enable password assigned or required.  This can also occur if the login has a privilege 15 assign locally or from the authentication server.  The current enable() function continues to send the secret password even when the prompt is in privilege exec mode already.  It was just changed to look for a the pattern (ie.  "ssword" before sending the secret password.

    def enable(self, cmd='', pattern='ssword', re_flags=re.IGNORECASE):
        """Enter enable mode."""
        output = ""
        msg = "Failed to enter enable mode. Please ensure you pass " \
              "the 'secret' argument to ConnectHandler."
        if not self.check_enable_mode():
            self.write_channel(self.normalize_cmd(cmd))
            try:
                output += self.read_until_prompt_or_pattern(pattern=pattern, re_flags=re_flags)
                """No enable or enable secret password"""
                if re.search(pattern, output):
                  self.write_channel(self.normalize_cmd(self.secret))
                  output += self.read_until_prompt()
            except NetMikoTimeoutException:
                raise ValueError(msg)
            if not self.check_enable_mode():
                raise ValueError(msg)
        return output